+++ package.json
@@ -29,6 +29,7 @@
     "passport-google-oauth": "^1.0.0",
     "passport-local": "^1.0.0",
     "socket.io": "^2.1.1",
+    "validate.js": "^0.12.0",
     "validator": "^10.7.1"
   },
   "devDependencies": {
