+++ public/css/_error.scss
@@ -1,0 +1,6 @@
+.error-message {
+  width: 100%;
+  margin-top: .25rem;
+  font-size: 80%;
+  color: #dc3545;
+}
