+++ public/css/_feed.scss
@@ -1,3 +1,37 @@
+
+.search-form {
+  margin: 0 auto;
+}
+
+.search-form-block {
+  padding-bottom: 30px;
+}
+
+.search-header-block {
+  display: block;
+  margin-top: 15px;
+  height: 40px;
+  border-top: 1px solid #e5e5e5;
+}
+
+.search-header {
+  text-align: center;
+}
+
+.search-header a {
+  padding: 10px 14px!important;
+}
+
+.container-form {
+  margin: 0 auto;
+  padding: 30px 0;
+  background-color: rgba(255,166,28,0.7);
+  border-radius: 28px;
+  box-shadow: 0 0 8px rgba(0,0,0,0.6);
+}
+
+
+
 .feed-box {
   margin: 0 auto;
 }
